linkerscript,memlog: massively shrink reserved regions
NPLL does *not* take that much memory for itself anymore now that it
features a proper allocator.  It's reserved region (dest reloc addr) can
be moved *far* higher and shrunken a lot.  This also requires moving the
memlog addr, which can realistically be shrunken as well given that it
gets disabled as soon as a real output device is added now.  These
should give far more space for kernels to be loaded.